// shared/contracts/database.types.ts
export type Database = {
  __InternalSupabase: {
    PostgrestVersion: '14.5';
  };
  public: {
    Tables: {
// ...
      categories: {
        Row: {
          color_hex: string;
          created_at: string;
          id: string;
          is_default: boolean;
          name: string;
          user_id: string | null;
        };
        Insert: {
          color_hex: string;
          created_at?: string;
          id?: string;
          is_default?: boolean;
          name: string;
          user_id?: string | null;
        };
        Update: {
          color_hex?: string;
          created_at?: string;
          id?: string;
          is_default?: boolean;
          name?: string;
          user_id?: string | null;
        };
        Relationships: [];
      };
// ...
      expenses: {
        Row: {
          amount: number;
          category_id: string;
          created_at: string;
          date: string;
          description: string | null;
          id: string;
          updated_at: string;
          user_id: string;
        };
        Insert: {
          amount: number;
          category_id: string;
          created_at?: string;
          date: string;
          description?: string | null;
          id?: string;
          updated_at?: string;
          user_id: string;
        };
        Update: {
          amount?: number;
          category_id?: string;
          created_at?: string;
          date?: string;
          description?: string | null;
          id?: string;
          updated_at?: string;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'expenses_category_id_fkey';
            columns: ['category_id'];
            isOneToOne: false;
            referencedRelation: 'categories';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
    };
// ...
  };
};